Move AuthLayout to auth feature components
- Relocate AuthLayout from navigation to auth feature components
- Update import paths in Login and Signup screens
- Remove deprecated AuthLayout file from navigation components
